+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ## [1.0.0] - 2023-02-17
-- Initial release
+- Initial release.
 
 ## [2.0.0] - 2023-02-17
-Used version 2.0 of package "sdm888/chuck-norris-jokes-demo"
+- Used version 2.0 of package "sdm888/chuck-norris-jokes-demo".
+
+## [2.1.0] - 2023-02-17
+- Added console command "chuck-norris-joke" for output jokes in console.